# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Port Daddy API
  version: 3.24.1
  description: >
    Authoritative port management and service orchestration for multi-agent
    development. Provides atomic port assignment, pub/sub messaging, distributed
    locks, agent registry, session coordination, and project:stack:context
    semantic identities.
  license:
    name: MIT
servers:
  - url: http://localhost:9876
    description: Local daemon default. Use `pd status` or `PORT_DADDY_URL` when your live daemon is elsewhere.
...
paths:
...
  /agents/{id}:
    get:
      tags: [agents]
      summary: Get agent info
      operationId: getAgent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent details
        '404':
          description: Agent not found
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [agents]
      summary: Unregister an agent
      operationId: unregisterAgent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent unregistered
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  unregistered:
                    type: boolean
        '400':
          description: Agent not found or error
        '500':
          $ref: '#/components/responses/InternalError'
...
  /changelog/agent/{agentId}:
    get:
      tags: [changelog]
      summary: List changelog entries for an agent
      operationId: listChangelogByAgent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 500
      responses:
        '200':
          description: Agent changelog entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  agentId:
                    type: string
                  entries:
                    type: array
                    items: {}
                  count:
                    type: integer
...
  /spawn/{id}:
    delete:
      tags: [spawn]
      summary: Kill a spawned agent
      operationId: killSpawnedAgent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent killed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  agentId:
                    type: string
                  message:
                    type: string
        '500':
          $ref: '#/components/responses/InternalError'

  # ===========================================================================
  # HARBORS
  # ===========================================================================
...
  /harbors/agent/{agentId}:
    get:
      tags: [harbors]
      summary: List harbors an agent is currently in
      operationId: listAgentHarbors
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent's harbors
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  harbors:
                    type: array
                    items: {}
                  count:
                    type: integer
                  agentId:
                    type: string
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
...
  responses:
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: internal server error